+++ backend/canisters/translations/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Start job to make pending payments each time a payment is added ([#5379](https://github.com/open-chat-labs/open-chat/pull/5379))
+- Trim proposed translations ([#5381](https://github.com/open-chat-labs/open-chat/pull/5381))
 
 ## [[2.0.1045](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1045-storage_index)] - 2024-02-05